% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtakeout-package.R
\docType{package}
\name{gtakeout-package}
\alias{gtakeout}
\alias{gtakeout-package}
\title{gtakeout: Extract data from Google Takeout}
\description{
gtakeout provides tools to extract, process, and manage data exported from Google Takeout. It simplifies the retrieval of structured information from various Google services, enabling users to analyze their personal data efficiently within R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jrosell/gtakeout}
  \item Report bugs at \url{https://github.com/jrosell/gtakeout/issues}
}

}
\author{
\strong{Maintainer}: Jordi Rosell \email{jroselln@gmail.com} (\href{https://orcid.org/0000-0002-4349-1458}{ORCID})

}
\keyword{internal}
