% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtakeout_reviews.R
\name{gtakeout_reviews}
\alias{gtakeout_reviews}
\title{Extract the reviews form a Takedown zip file.
#'}
\usage{
gtakeout_reviews(zipfile, ..., exdir = here::here("data"))
}
\arguments{
\item{zipfile}{Required path to the zip file containing the reviews}

\item{...}{Other optional arguments}

\item{exdir}{Optional directory to extract the contents of the zip file}
}
\value{
A data frame of reviews
}
\description{
Extract the reviews form a Takedown zip file.
#'
}
\examples{
zip_path <- system.file(
    "extdata",
    "takeout-example.zip",
    package = "gtakeout"
)
gtakeout_reviews(zip_path, exdir = tempdir())

}
