% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtakeout-package.R
\docType{package}
\name{gtakeout-package}
\alias{gtakeout}
\alias{gtakeout-package}
\title{gtakeout: Extract Data from Google Takeout}
\description{
Provides functions to analyze data exported from 'Google Takeout'. The package supports unzipping archives and extracting user review data from Google Business Profile exports into tidy data frames for further analysis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jrosell/gtakeout}
  \item Report bugs at \url{https://github.com/jrosell/gtakeout/issues}
}

}
\author{
\strong{Maintainer}: Jordi Rosell \email{jroselln@gmail.com} (\href{https://orcid.org/0000-0002-4349-1458}{ORCID}) [copyright holder]

}
\keyword{internal}
