% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtakeout_reviews.R
\name{gtakeout_reviews}
\alias{gtakeout_reviews}
\title{Extract the reviews form a Takedown zip file.
#'}
\usage{
gtakeout_reviews(zipfile, ..., exdir = here::here("data"))
}
\arguments{
\item{zipfile}{Required path to the zip file containing the reviews}

\item{...}{Additional arguments (if applicable)}

\item{exdir}{Optional directory to extract the contents of the zip file}
}
\value{
A data frame of reviews
}
\description{
Extract the reviews form a Takedown zip file.
#'
}
